requirements.txt
Flask==1.1.2,
google-cloud-datastore==1.12.0,
Flask_Cors==3.0.8,